Revert "Added batching using DB. (#724)"
This reverts commit f4fcc03ee75ccb256aa6e46e4eb299497797fc74.
