+++ src/task_monitor/tasks/mod.rs
@@ -1,6 +1,5 @@
-pub mod create_batches;
 pub mod delete_identities;
 pub mod finalize_identities;
 pub mod insert_identities;
 pub mod monitor_txs;
-pub mod process_batches;
+pub mod process_identities;
